Remove redundant app layer
After: `Application` → `Router`
Before: `Application` → `App` → `Router` (redundant layer)
