+++ apps/web/core/application.rb
@@ -3,7 +3,7 @@
 require 'onetime/application'
 require 'onetime/middleware'
 
-require_relative 'app'
+require_relative 'controllers'
 
 module Core
   class Application < Onetime::Application::Base
@@ -37,8 +37,22 @@ module Core
     protected
 
     def build_router
-      # Return the Core app instance
-      Core::App.new
+      routes_path = File.join(Onetime::HOME, 'apps/web/core/routes')
+      router      = Otto.new(routes_path)
+
+      # Enable CSP nonce support for enhanced security
+      router.enable_csp_with_nonce!(debug: OT.debug?)
+
+      # Register authentication strategies for Web Core
+      require_relative 'auth_strategies'
+      Core::AuthStrategies.register_all(router)
+
+      # Default error responses
+      headers             = { 'content-type' => 'text/html' }
+      router.not_found    = [404, headers, ['Not Found']]
+      router.server_error = [500, headers, ['Internal Server Error']]
+
+      router
     end
   end
 end
